Finish changing asset managing code.
