+++ game/__init__.py
@@ -1,8 +1,7 @@
 from .settings import width, height, screen_dims
 
-from .screens.levels.level import Level
+from .assets import Assets
 
+from .screens.levels.level import Level
 from .screens.menus.main_menu import MainMenu
 from .screens.menus.game_over import GameOver
-
-from .asset_manager import AssetManager
