+++ game/images/enemy_ship.png
