+++ game/images/player_ship.png
